Remove compatibility with sbt 0.13
